+++ CHANGES.md
@@ -16,6 +16,8 @@ Changes
 - ScopeBuilderVisitor now also tracks usage
 - Fix ScopeBuilderVisitor uncorrectly marking ScopeEntry as GLOBAL
 - Fix locating magik-lint.properties in products
+- Make parsing/checking files parallel in magik-lint
+- Check Formatting now requires empty line after TRANSMIT
 
 
 0.3.2 (2019-10-29)